pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "taskgen-ai"
version = "3.2.0"
authors = [
  { name="John Tan Chong Min", email="tanchongmin@gmail.com" },
]
description = "A Task-based agentic framework building on StrictJSON outputs by LLM agents"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["openai>=1.3.6", "dill>=0.3.7", "termcolor>=2.3.0", "requests", "langchain",
"PyPDF2", "python-docx", "pandas", "xlrd"]

[project.urls]
Homepage = "https://github.com/simbianai/taskgen"
Issues = "https://github.com/simbianai/taskgen/issues"